streamlit/requirements.txt
streamlit
pandas==1.3.3
protobuf==3.20.1
numpy==1.21.2
boto3
PyPDF2
openai==0.28.0
altair==4.1.0
requests
python-dotenv

